#int -> double 2

+++ src/main/java/cn/edu/nju/modules/rank/SeedsManager.java
@@ -9,7 +9,7 @@ public interface SeedsManager {
     void sort(); // 对种子进行排序
 
     @UnmodifiableSignature
-    void addSeed(String seedPath, int coverage,int energy);
+    void addSeed(String seedPath, double coverage,double energy);
 
     @UnmodifiableSignature
     void register(ResourcesManager resourcesManager); // implement class have this resource manager
